Added Order and Product exceptions and more unit tests for the exceptions for the Order manager and the Product manager

+++ individual-sem3-backend/src/main/java/individual/individualsem3backend/business/impl/OrderManagerImpl.java
@@ -1,6 +1,7 @@
 package individual.individualsem3backend.business.impl;
 
 import individual.individualsem3backend.business.OrderManagerUseCase;
+import individual.individualsem3backend.business.exception.OrderException;
 import individual.individualsem3backend.domain.Order;
 import individual.individualsem3backend.persistence.OrderRepository;
 import lombok.AllArgsConstructor;
@@ -16,33 +17,64 @@ public class OrderManagerImpl implements OrderManagerUseCase {
     private OrderRepository orderRepository;
 
     public List<Order> getAllOrders(Integer userId){
-        return orderRepository.findAll(userId);
+        if(userId > -1){
+            return orderRepository.findAll(userId);
+        }
+        else{
+            throw new OrderException("Cannot find orders with a negative user id.");
+        }
     }
 
     public Optional<Order> findOrderById(Integer orderId){
-        return Optional.ofNullable(orderRepository.findOrderById(orderId));
+        if(orderId > -1){
+            return Optional.ofNullable(orderRepository.findOrderById(orderId));
+        }
+        else{
+            throw new OrderException("Cannot find orders with a negative user id.");
+        }
     }
     public Order create(Order neworder){
-        Order order = Order.builder().userId(neworder.getUserId()).products(neworder.getProducts())
-                .bundle_or_not(neworder.isBundle_or_not())
-                .date_of_purchase(neworder.getDate_of_purchase()).build();
 
-        return orderRepository.save(order);
+        if(neworder != null){
+            Order order = Order.builder().userId(neworder.getUserId()).products(neworder.getProducts())
+                    .bundle_or_not(neworder.isBundle_or_not())
+                    .date_of_purchase(neworder.getDate_of_purchase()).build();
+
+            return orderRepository.save(order);
+        }
+        else{
+            throw new OrderException("Something went wrong while creating the order.");
+        }
     }
 
     public void delete(Integer orderId){
-        orderRepository.deleteById(orderId);
+        if(orderId > -1){
+            orderRepository.deleteById(orderId);
+        }
+        else{
+            throw new OrderException("Cannot delete order with a negative id.");
+        }
     }
 
     public Order update(Order updatedOrder){
-        Order oldOrder = orderRepository.findOrderById(updatedOrder.getId());
+        try{
+            if(updatedOrder.getId() > -1 && updatedOrder != null){
+                Order oldOrder = orderRepository.findOrderById(updatedOrder.getId());
 
-        oldOrder.setProducts(updatedOrder.getProducts());
-        oldOrder.setBundle_or_not(updatedOrder.isBundle_or_not());
-        oldOrder.setDate_of_purchase(updatedOrder.getDate_of_purchase());
+                oldOrder.setProducts(updatedOrder.getProducts());
+                oldOrder.setBundle_or_not(updatedOrder.isBundle_or_not());
+                oldOrder.setDate_of_purchase(updatedOrder.getDate_of_purchase());
 
-        orderRepository.update(oldOrder);
+                orderRepository.update(oldOrder);
 
-        return oldOrder;
+                return oldOrder;
+            }
+            else{
+                throw new OrderException("Something went wrong while updating the order.");
+            }
+        }
+        catch(Exception ex){
+            throw new OrderException("Something went wrong while updating the order.");
+        }
     }
 }
